Remove stall tests.
