+++ tests/PropertiumUnitTests/SearchProperties/Get/Timeout_.cs
@@ -6,7 +6,6 @@ using Propertium.UnitTests.DataAttributes;
 using Shouldly;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
 using TimeoutEx;
 using Xunit;
 
@@ -15,73 +14,6 @@ namespace Propertium.UnitTests.SearchProperties.Get
     public class Timeout_
     {
         public class Given_3_webElements_with_same_locator_properties_
-        {
-            [Theory]
-            [InlineAutoMoqData(0)]
-            [InlineAutoMoqData(1)]
-            [InlineAutoMoqData(2)]
-            public void When_index_is_between_0_and_2_Then_returns_webElement_at_matching_index(
-                int index,
-                IFindsByFluentSelector<IWebElement> webDriver,
-                [Frozen]IReadOnlyCollection<IWebElement> webElements,
-                string locator,
-                string value
-                )
-            {
-                // Arrange
-                var timeout = 100.Milliseconds();
-                var expected = webElements.ElementAt(index);
-                Mock.Get(webDriver).Setup(x => x.FindElements(locator, value)).Returns(webElements);
-                var sut = new SearchProperty<IWebElement>(locator, value, webDriver, index, timeout);
-
-                // Act
-                var actual = sut.Get(timeout);
-
-                // Assert
-                actual.ShouldBe(expected);
-            }
-
-            [Theory, AutoMoqData]
-            public void When_index_is_out_of_range_Then_returns_null(
-                IFindsByFluentSelector<IWebElement> webDriver,
-                [Frozen]IReadOnlyCollection<IWebElement> webElements,
-                string locator,
-                string value)
-            {
-                // Arrange
-                var timeout = 100.Milliseconds();
-                Mock.Get(webDriver).Setup(x => x.FindElements(locator, value)).Returns(webElements);
-                int indexOutOfRange = webElements.Count + 1;
-                var sut = new SearchProperty<IWebElement>(locator, value, webDriver, indexOutOfRange, timeout);
-
-                // Act
-                var actual = sut.Get(timeout);
-
-                // Assert
-                actual.ShouldBeNull();
-            }
-
-            [Theory, AutoMoqData]
-            public void When_index_is_not_defined_Then_returns_first_WebElement(
-                IFindsByFluentSelector<IWebElement> webDriver,
-                [Frozen]IReadOnlyCollection<IWebElement> webElements,
-                string locator,
-                string value)
-            {
-                // Arrange
-                var timeout = 100.Milliseconds();
-                Mock.Get(webDriver).Setup(x => x.FindElements(locator, value)).Returns(webElements);
-                var sut = new SearchProperty<IWebElement>(locator, value, webDriver, timeout);
-
-                // Act
-                var actual = sut.Get(timeout);
-
-                // Assert
-                actual.ShouldBe(webElements.First());
-            }
-        }
-
-        public class Given_a_defaultCancellationToken_And_3_webElements_with_same_locator_properties_
         {
             [Theory]
             [InlineAutoMoqData(0)]
@@ -138,8 +70,7 @@ namespace Propertium.UnitTests.SearchProperties.Get
                 // Arrange
                 var timeout = 100.Milliseconds();
                 Mock.Get(webDriver).Setup(x => x.FindElements(locator, value)).Returns(webElements);
-                int indexOutOfRange = webElements.Count + 1;
-                var sut = new SearchProperty<IWebElement>(locator, value, webDriver, 150.Milliseconds());
+                var sut = new SearchProperty<IWebElement>(locator, value, webDriver, 200.Milliseconds());
 
                 // Act
                 var actual = sut.Get(timeout);
